requirements.txt
ansible==12.0.0
ansible-core==2.19.2
cffi==2.0.0
cryptography==46.0.1
et_xmlfile==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==25.0
pycparser==2.23
PyYAML==6.0.2
resolvelib==1.2.0
six==1.17.0
